Added the first working unit test for the Smil data structure.
The test flushed out some bugs and mistakes in my implementation of SMIL 10 Specification - good news. 

We need more tests for the smil data structure e.g. to make sure our understanding of what we need to implement works for my collection of smil files. We also need to implement support for extracting the text for the <text> elements.